+++ Enemies/Goriya.cs
@@ -170,10 +170,11 @@ namespace LegendOfZelda
         public void StopFlashing()
         {
             Sprites[CurrentSprite].flashing = false;
-            public void DropItem()
-            {
-                Center = EnemyUtilities.GetCenter(Position, 16, 16);
-                EnemyItemDrop.DropClassBItem(Center);
-            }
+        }
+        public void DropItem()
+        {
+            Center = EnemyUtilities.GetCenter(Position, 16, 16);
+            EnemyItemDrop.DropClassBItem(Center);
         }
     }
+}
